pyproject.toml
[project]
name = "marco-chrissy-site"
version = "0.1.0"
description = "Personal website with photo galleries"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pelican>=4.9.0",
    "markdown>=3.5.0",
    "pillow>=10.2.0",  # For image processing
    "click>=8.1.0",    # For CLI
    "jinja2>=3.1.0",   # For templates
    "jsonschema>=4.0.0",  # For config validation
    "normpic @ git+https://github.com/marcus-grant/normpic.git",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
]
test = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "requests>=2.31.0",  # For serve command E2E tests
    "beautifulsoup4>=4.12.0",  # For HTML parsing in E2E tests
]
deploy = [
    "requests>=2.31.0",  # For CDN API calls
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.hatch.metadata]
allow-direct-references = true

[project.scripts]
site = "cli.main:main"
build = "orchestrator.builder:main"
deploy = "deployer.cdn:main"
galleria = "galleria.__main__:main"

[tool.ruff]
line-length = 88
target-version = "py312"
lint.select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]
lint.ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.per-file-ignores]
"test/**" = ["S101"]  # Allow assert in tests

[tool.pytest.ini_options]
testpaths = [
    "test",
    "galleria/test",
]
pythonpath = ["."]
addopts = "--verbose"

[tool.coverage.run]
source = ["galleria", "orchestrator", "deployer"]
omit = [
    "*/test/*",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "raise AssertionError",
    "raise NotImplementedError",
]

[dependency-groups]
dev = [
    "pyfakefs>=5.10.2",
    "pytest>=9.0.1",
    "watchdog>=3.0.0",  # For file watching in serve command
]
